Working Pie Chart with Dashboard Context, also separated in different files PieChart's functions
