+++ src/components/Dashboard/DashboardHeader/DashboardHeader.module.css
@@ -9,10 +9,12 @@ font-size: 28px;}
 .container{
   width: 100%;
   height: 100%;
+  max-height: 80px;
   display: flex;
   flex-direction: row;
   justify-content: flex-start;
-  align-items: center;
+  align-items:flex-start;
+  /* background-color: red; */
 }
 
 .wrapperTitle{
@@ -35,4 +37,8 @@ font-size: 28px;}
   justify-content: flex-start;
   align-items: center;
   margin-left: 20px;
+}
+
+.MonthFees{
+  background-color: red;
 }
