+++ src/components/Dashboard/MonthFees/MonthFees.module.css
@@ -1,25 +1,43 @@
 .container{
-  display: flex;
+  /* display: flex;
   flex-direction: column;
-  width: 93%;
-  height: 100% !important; 
-  border: 2px solid #eceeF6;
+  width: 100%;
+  height: 100%; 
+  border: 2px solid #eceeF6; */
+  /* background-color: rebeccapurple; */
   /*CORRECT COLOR: #ECEEF6*/
-  border-radius: 10px;
+  /* border-radius: 10px;
   outline: none;
   margin: 20px;
   padding: 4px;
   display: flex;
   align-items: center;
+  justify-content: center; */
+  background-color: rgba(255, 0, 0, 0.15);
+  display: flex;
+  flex-direction: column;
+  align-items: center;
   justify-content: center;
+  width: 100%;
+  height: 100%;
+  padding: 5px;
+}
+.wrapper{
+  width: 95%;
+  background-color: rgba(0, 0, 255, 0.312);
+  height: 95%;
+  border: 2px solid #eceeF6;
 }
 .titleWrapper{
   display: flex;
   width: 100%;
+  height: auto;
+  background-color: blanchedalmond;
 }
 .title{
   font-size: 16px;
   width: 100%;
+  height: 100%;
   margin-top: 35px;
   margin-left: 20px;
