+++ src/components/Dashboard/MonthFees/MonthFeesPieChart/data.js
@@ -1,0 +1,4 @@
+// const data = [
+//   { name: "Pagado", value: 270.30 },
+//   { name: "Gastos", value: 923.28 }
+// ];
